@@ -4,10 +4,7 @@ import ExpenseDate from "./ExpenseDate";
 import "./ExpenseItem.css";
 
 function ExpenseItem(props) {
-  const [title, setTitle] = useState(props.title);
-
   const clickHandler = () => {
-    setTitle("Updated");
     console.log("Clicked!!!");
   };
 
@@ -15,7 +12,7 @@ function ExpenseItem(props) {
     <Card className="expense-item">
       <ExpenseDate date={props.date} />
       <div className="expense-item__description">
-        <h2>{title}</h2>
+        <h2>{props.title}</h2>
         <div className="expense-item__price">${props.amount}</div>
       </div>
       <button onClick={clickHandler}>Change Title</button>
